@@ -9,14 +9,12 @@
   "dependencies": {
     "@types/jest": "^27.0.2",
     "@types/node": "^16.11.6",
-    "@types/react": "^17.0.33",
-    "@types/react-dom": "^17.0.10",
     "axios": "^0.21.1",
     "connected-react-router": "^6.9.1",
     "formik": "^2.2.6",
     "history": "4.10.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-hook-form": "^7.12.2",
     "react-redux": "^7.2.4",
     "react-router-dom": "^5.2.0",
@@ -44,6 +42,8 @@
   ],
   "devDependencies": {
     "@types/history": "^4.7.9",
+    "@types/react": "^17.0.43",
+    "@types/react-dom": "^17.0.14",
     "@types/react-router-dom": "^5.3.2",
     "@types/redux-logger": "^3.0.9",
     "react-scripts": "^4.0.3"
