@@ -1,29 +1,23 @@
-import { useState } from "react";
-import { Redirect, useParams } from "react-router-dom";
+import { useParams } from "react-router-dom";
 import { attemptGetConfirmation } from "../store/thunks/auth";
 import { Error } from "../components";
-import { useAppSelector, useAppDispatch } from "src/store/hooks";
+import { useAppDispatch } from "src/store/hooks";
+import { useServerError } from "src/hooks/useServerError";
 
 export default function RegisterConfirmationPage() {
-  const { isAuth } = useAppSelector((state) => state.user);
-  const [serverError, setServerError] = useState(null);
   const dispatch = useAppDispatch();
+  const { serverError, handleServerError } = useServerError();
+
   const { token } = useParams<{ token: string }>();
 
-  function doSubmit() {
-    dispatch(attemptGetConfirmation(token)).catch((error) => {
-      if (error.response) {
-        setServerError(error.response.data.message);
-      }
-    });
-  }
+  const handleSubmit = () => {
+    dispatch(attemptGetConfirmation(token)).catch(handleServerError);
+  };
 
-  return isAuth ? (
-    <Redirect to='/home' />
-  ) : (
+  return (
     <div className='container'>
       <p>Click here to confirm your email</p>
-      <button onClick={doSubmit}>Confirmation</button>
+      <button onClick={handleSubmit}>Confirmation</button>
       {serverError && <Error>{serverError}</Error>}
     </div>
   );
