@@ -1,9 +1,11 @@
 import axios from "axios";
 
-axios.defaults.baseURL = `${process.env.REACT_APP_API_URL}/api`;
-axios.defaults.withCredentials = true;
+const axiosInstance = axios.create({
+  baseURL: `${process.env.REACT_APP_API_URL}/api`,
+  withCredentials: true,
+});
 
-axios.interceptors.response.use(
+axiosInstance.interceptors.response.use(
   (res) => res,
   (error) => {
     return Promise.reject(error);
@@ -11,10 +13,10 @@ axios.interceptors.response.use(
 );
 
 const httpService = {
-  get: axios.get,
-  post: axios.post,
-  put: axios.put,
-  delete: axios.delete,
+  get: axiosInstance.get,
+  post: axiosInstance.post,
+  put: axiosInstance.put,
+  delete: axiosInstance.delete,
 };
 
 export default httpService;
